[content script] time as number and global
I am going to store the time as a Number, which makes things easier. I'm also going to have a global CHANNEL_TIME, which cuts down on the calls to storage.  I've also got a global INTERVAL_SECONDS, which I'm using (1) to add to CHANNEL_TIME and (2) to set setInterval()