[popup] basic display
Shows the watched channels in a table in the popup. Currently, I cutoff the channels to show at 30 minutes. I might want to let it be customizable later, but it seemed like a decent default.

I think I want to redo the formatting on the injected div for the time display on the Twitch page. Maybe just dd:hh:mm