channel deletion
Added in the ability to delete a channel, effectively resetting it.

This was made in response to a request by a reviewer. An alternative option was to add in time-series functionality, but (1) that was was more work, and (2) I myself want this feature so that I can clean up the list.